Add files via upload
In fibration attention, a lifting simultaneously updates a query and abductively finds a key to explain a value.  This all depends on the key-value pairing (k,v). By determining other keys that are highly similar to k, to also explain v, we can enhance key-value semantics to achieve more meaningful attention. Moreover, we can recursively layer keys upon keys in a tree, to deepen the set of value explanations.