Add files via upload
In a LLM database retrieval paradigm, Q&A seeks a semantic handshake via intermediary keys.  A key-value pairing p: K -> V gives a value v in V semantics via its inverse image under p, like looking up the definition of a word in a dictionary.  Keys can also be used to give meaning to queries via a neural network d: K -> Q.  Note that semantics p and d are neural networks to be learned.  Moreover, we can learn a fibration lifting τ that connects p and d in a semantic handshake, facilitating meaningful answers to queries.  Technically, d and p are Chu adjoints under τ, and the value returned to a query q is approximately p(c(q)) ε V, where c: Q -> K is a section of d.