Add files via upload
Given a token in a context string, an implicit deep learning model reveals a latent token, that we dub a ‘noken’, and token representations, such that keys embue meaning to queries and values.  This construction can then be construed as a machine oracle that aligns meaning utilizing a universal interlingua for next-token and sequence-to-sequence learning.